Moved more stuff into the scripts folder
