+++ scripts/card_updater/index.js
@@ -1,11 +1,11 @@
-const vanillaCards = require("../card_creator/vanilla/.ignore.cards.json");
-const { Game } = require("../src/game");
+const vanillaCards = require("../../card_creator/vanilla/.ignore.cards.json");
+const { Game } = require("../../src/game");
 
 const game = new Game({}, {});
 game.dirname = __dirname + "/../";
 
-game.functions.importCards("../cards");
-game.functions.importConfig("../config");
+game.functions.importCards(__dirname + "/../../cards");
+game.functions.importConfig(__dirname + "/../../config");
 
 let customCards = game.functions.getCards(false);
