+++ scripts/card_updater/run.bat
