+++ scripts/card_updater/run.sh
